Add note for creating a new ECC Ed29919 key (#3)
Add notes and code for creating a whole new ECC Ed29919 OpenPGP key with GnuPG.

ECC keys are modern, and curve Ed25519 is considered to be secure at this time. Subkeys
can be generated from a primary key in an encrypted volume, without the certifying
primary key ever leaving the encrypted volume. The subkeys can be distributed for use
with a shorter passphrase for ease of use.